+++ src/test/java/cart/service/ProductServiceImplTest.java
@@ -12,7 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
-import cart.controller.request.ProductUpdateRequest;
+import cart.service.dto.ProductUpdateRequest;
 import cart.domain.Product;
 import cart.controller.response.ProductResponse;
 import cart.repository.ProductRepository;
